+++ compiler-core/src/exhaustiveness.rs
@@ -3408,6 +3408,7 @@ impl ConstructorSpecialiser {
             name,
             parameters,
             documentation,
+            deprecation,
         } = v;
         let parameters = parameters
             .iter()
@@ -3421,6 +3422,7 @@ impl ConstructorSpecialiser {
             name: name.clone(),
             parameters,
             documentation: documentation.clone(),
+            deprecation: deprecation.clone(),
         }
     }